[item] Separate touch event processing from UI code. Fixes JB#45998
Touch event processing in minui originally supported only multitouch
protocol B and support for protocol A was added later on. There are some
more or less subtle differences in how various devices implement A or B
protocols and dealing with these is difficult as the event processing code
is littered with heuristic conditionals and interleaved with UI level code.

Move low level touch processing to a separate class. Probe input device
capabilities to find out whether it uses protocol A or B. Use event
processing state machine that is appropriate for the protocol and expose
finger pressed, moved and lifted events for consumption at UI level.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>